refactor: split content search result rows
